+++ src/main/java/kitchenpos/menu/domain/repository/MenuProductRepository.java
@@ -1,10 +1,7 @@
-package kitchenpos.menu.repository;
+package kitchenpos.menu.domain.repository;
 
 import kitchenpos.menu.domain.MenuProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
 public interface MenuProductRepository extends JpaRepository<MenuProduct, Long> {
-    List<MenuProduct> findAllByMenuId(final Long menuId);
 }
