+++ src/main/java/kitchenpos/menu/domain/repository/MenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.repository;
+package kitchenpos.menu.domain.repository;
 
 import kitchenpos.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,7 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
-    long countByIdIn(final List<Long> ids);
-
     List<Menu> findByIdIn(List<Long> menuIds);
 }
