+++ src/main/java/kitchenpos/menugroup/application/MenuGroupService.java
@@ -3,7 +3,7 @@ package kitchenpos.menugroup.application;
 import kitchenpos.menugroup.application.dto.MenuGroupCreateRequest;
 import kitchenpos.menugroup.application.dto.MenuGroupResponse;
 import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.menugroup.repository.MenuGroupRepository;
+import kitchenpos.menugroup.domain.repository.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
