+++ src/main/java/kitchenpos/order/domain/repository/OrderLineItemRepository.java
@@ -1,10 +1,7 @@
-package kitchenpos.order.repository;
+package kitchenpos.order.domain.repository;
 
 import kitchenpos.order.domain.OrderLineItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
 public interface OrderLineItemRepository extends JpaRepository<OrderLineItem,Long> {
-    List<OrderLineItem> findAllByOrderId(final Long orderId);
 }
