+++ src/main/java/kitchenpos/order/domain/repository/OrderRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.repository;
+package kitchenpos.order.domain.repository;
 
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
