+++ src/main/java/kitchenpos/ordertable/domain/repository/OrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.repository;
+package kitchenpos.ordertable.domain.repository;
 
 import kitchenpos.ordertable.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
