+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -3,7 +3,7 @@ package kitchenpos.product.application;
 import kitchenpos.product.application.dto.ProductCreateRequest;
 import kitchenpos.product.application.dto.ProductResponse;
 import kitchenpos.product.domain.Product;
-import kitchenpos.product.repository.ProductRepository;
+import kitchenpos.product.domain.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
