+++ src/main/java/kitchenpos/product/domain/repository/ProductRepository.java
@@ -1,10 +1,7 @@
-package kitchenpos.product.repository;
+package kitchenpos.product.domain.repository;
 
 import kitchenpos.product.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
 public interface ProductRepository extends JpaRepository<Product, Long> {
-    List<Product> findByIdIn(final List<Long> menuProductIds);
 }
